Update Entities and EntitiesMP.

+++ SamTSE/Sources/EntitiesMP/Common/Particles.cpp
@@ -1045,7 +1045,7 @@ void Particles_BeastProjectileTrail( CEntity *pen, FLOAT fSize, FLOAT fHeight, I
       vZ*(afStarsPositions[iStar][2]*time*fSize*1.5);
 
     FLOAT3D vPos = GET_POS( fT);
-    COLOR colStar = pTD->GetTexel( ClampUp(FloatToInt(fT*8192),8191), 0);
+    COLOR colStar = pTD->GetTexel( ClampUp((int)FloatToInt(fT*8192),8191), 0);
 
     if( fT>BEAST_PROJECTILE_LINE_PARTICLES)
     {
@@ -1107,7 +1107,7 @@ void Particles_BeastBigProjectileTrail( CEntity *pen, FLOAT fSize, FLOAT fZOffse
       vZ*(afStarsPositions[iStar][2]*time*fSize*1.5);
 
     FLOAT3D vPos = GET_POS_BIG( fT);
-    COLOR colStar = pTD->GetTexel( ClampUp(FloatToInt(fT*8192),8191), 0);
+    COLOR colStar = pTD->GetTexel( ClampUp((int)FloatToInt(fT*8192),8191), 0);
 
     if( fT>BIG_BEAST_PROJECTILE_LINE_PARTICLES)
     {
@@ -3494,7 +3494,7 @@ void Particles_LavaFlow( CEntity *pen, FLOAT fStretchAll, FLOAT fSize, FLOAT fHe
         vY*(fT*fT*-4.0f+(afStarsPositions[iStar][1]*fPowerFactor*0.1)) +
         vZ*(afStarsPositions[iStar][2]*fPowerFactor*fT*fStretchAll);
     
-      COLOR colLava = pTD->GetTexel( ClampUp(FloatToInt(fT*2048),2047), 0);
+      COLOR colLava = pTD->GetTexel( ClampUp((int)FloatToInt(fT*2048),2047), 0);
       ULONG ulA = FloatToInt( ((colLava&CT_AMASK)>>CT_ASHIFT) * fFade);
       colLava = (colLava&~CT_AMASK) | (ulA<<CT_ASHIFT);
       Particle_RenderSquare( vPos, fSize, 0, colLava);
